pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "llm-text-to-sql-finance"
version = "0.1.0"
description = "Governed Text-to-SQL interface for finance teams with RBAC, audit logging, and KPI glossary"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.0",
    "snowflake-connector-python>=3.6",
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
    "httpx>=0.27",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.4",
]

[tool.ruff]
target-version = "py311"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"

[tool.mypy]
python_version = "3.11"
strict = true
